Changes on the reports

+++ ProyectoFinalAp2/Models/EntradaProductos.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -13,12 +14,10 @@ namespace ProyectoFinalAp2.Models
         public int EntradaProductoId { get; set; }
         public int UsuarioId { get; set; }
         public DateTime Fecha { get; set; }
-        public int CantidadTotal { get; set; }
+        public decimal CantidadTotal { get; set; }
         public virtual List<DetalleEntradaProductos> DetalleEntrada { get; set; }
         //[ForeignKey("UsuarioId")]
         //public virtual Usuarios Usuarios { get; set; }
-        //[ForeignKey("ProductoId")]
-        //public virtual Productos Productos { get; set; }
 
         public EntradaProductos()
         {
